README
This java/xml  android app was created for the purpose of scrolling language learning cards horizontally.  You can add as many cards as you like. The cards are scrollable horizontally, and the text is scrollable vertically.  
The app was built on an android tablet using AIDE, with android 4.2.2,  and tested on same android tablet, and also tested on a smartphone with android version 6